requirements.txt
polars>=1.0.0
numpy>=1.24.0
pyarrow>=14.0.0

scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

torch>=2.0.0

plotly>=5.18.0

tqdm>=4.66.0
pyyaml>=6.0.0
joblib>=1.3.0

pytest>=7.4.0
